refact command

+++ src/Crafthoe.Protocol/Commands/ChunkUpdateCommand.cs
@@ -3,7 +3,7 @@ namespace Crafthoe.Protocol;
 [StructLayout(LayoutKind.Sequential, Pack = 1)]
 public struct ChunkUpdateCommand : ICommand
 {
-    public static int CommandId => (int)Commands.ChunkUpdate;
+    public static ushort CommandId => (ushort)Commands.ChunkUpdate;
 
     public Vector2i Cloc;
 }
