+++ src/Crafthoe.Protocol/Commands/ForgetChunkCommand.cs
@@ -3,7 +3,7 @@ namespace Crafthoe.Protocol;
 [StructLayout(LayoutKind.Sequential, Pack = 1)]
 public struct ForgetChunkCommand : ICommand
 {
-    public static int CommandId => (int)Commands.ForgetChunk;
+    public static ushort CommandId => (ushort)Commands.ForgetChunk;
 
     public Vector2i Cloc;
 }
