+++ src/Crafthoe.Protocol/Commands/MovePlayerCommand.cs
@@ -3,7 +3,7 @@ namespace Crafthoe.Protocol;
 [StructLayout(LayoutKind.Sequential, Pack = 1)]
 public struct MovePlayerCommand : ICommand
 {
-    public static int CommandId => (int)Commands.MovePlayer;
+    public static ushort CommandId => (ushort)Commands.MovePlayer;
 
     public MovementStep Step;
 }
