+++ src/Crafthoe.Protocol/Commands/PingCommand.cs
@@ -3,7 +3,7 @@ namespace Crafthoe.Protocol;
 [StructLayout(LayoutKind.Sequential, Pack = 1)]
 public struct PingCommand : ICommand
 {
-    public static int CommandId => (int)Commands.Ping;
+    public static ushort CommandId => (ushort)Commands.Ping;
 
     public long Timestamp;
     public long Id;
