+++ src/Crafthoe.Protocol/Commands/PongCommand.cs
@@ -3,7 +3,7 @@ namespace Crafthoe.Protocol;
 [StructLayout(LayoutKind.Sequential, Pack = 1)]
 public struct PongCommand : ICommand
 {
-    public static int CommandId => (int)Commands.Pong;
+    public static ushort CommandId => (ushort)Commands.Pong;
 
     public PingCommand Ping;
 }
