+++ src/Crafthoe.Protocol/Commands/SpawnPlayerCommand.cs
@@ -3,5 +3,5 @@ namespace Crafthoe.Protocol;
 [StructLayout(LayoutKind.Sequential, Pack = 1)]
 public struct SpawnPlayerCommand : ICommand
 {
-    public static int CommandId => (int)Commands.SpawnPlayer;
+    public static ushort CommandId => (ushort)Commands.SpawnPlayer;
 }
