+++ src/Crafthoe.Protocol/Commands/WorldIndicesUpdateCommand.cs
@@ -3,5 +3,5 @@ namespace Crafthoe.Protocol;
 [StructLayout(LayoutKind.Sequential, Pack = 1)]
 public struct WorldIndicesUpdateCommand : ICommand
 {
-    public static int CommandId => (int)Commands.WorldIndicesUpdate;
+    public static ushort CommandId => (ushort)Commands.WorldIndicesUpdate;
 }
